+++ babel.config.js
@@ -1,16 +1,3 @@
 module.exports = {
-    env: {
-        development: {
-            sourceMaps: true,
-            plugins: ['source-map-support'],
-        }
-    },
-
-    plugins: [
-        'codegen'
-    ],
-
-    presets: [
-        'bili/babel',
-    ],
+    plugins: ['codegen'],
 }