+++ test/parser/valid.js
@@ -3,8 +3,6 @@ const Util      = require('util')
 const { parse } = require('../..')
 const VALID     = require('./fixtures/_valid.js')
 
-require('array-flat-polyfill') // for Node v10
-
 for (const [path, want] of Object.entries(VALID)) {
     const name = Util.format('%j -> %j', path, want)